refactor: rename component from source to mapping
The old component `source` was actually a wrapper around source maps. Let's use `mapping` which we already uses for the instances of that component. The real motivation for that change is to use `source` for actual sources in the future.
