+++ components/instrumentation/default/index.test.mjs
@@ -1,5 +1,5 @@
 import { assertThrow, assertDeepEqual } from "../../__fixture__.mjs";
-import { createMirrorSourceMap } from "../../source/index.mjs";
+import { createMirrorSourceMap } from "../../mapping/index.mjs";
 import {
   createConfiguration,
   extendConfiguration,
