+++ components/instrumentation/default/visit.mjs
@@ -8,7 +8,7 @@ import {
   hasOwnProperty,
   coalesce,
 } from "../../util/index.mjs";
-import { mapSource } from "../../source/index.mjs";
+import { mapSource } from "../../mapping/index.mjs";
 import { logDebugWhen, logErrorWhen } from "../../log/index.mjs";
 import { stringifyLocation } from "../../location/index.mjs";
