+++ components/instrumentation/default/visit.test.mjs
@@ -1,7 +1,7 @@
 import { assertEqual } from "../../__fixture__.mjs";
 import { stringifyLocation } from "../../location/index.mjs";
 import { createCounter } from "../../util/index.mjs";
-import { createMirrorSourceMap } from "../../source/index.mjs";
+import { createMirrorSourceMap } from "../../mapping/index.mjs";
 import { hashFile } from "../../hash/index.mjs";
 import { normalize, parse, generate } from "./__fixture__.mjs";
 import { visit } from "./visit.mjs";
